Create words.html
Write a HTML page that has one input, which can take multi-line text and a submit button. Once the user clicks the submit button, it should show the     number of characters, words and lines in the text entered using an alert message. Words are separated with white space and lines are separated with new line character.